requirements.txt
asn1crypto==0.24.0
backcall==0.1.0
bcrypt==3.1.6
cffi==1.12.2
cryptography==2.6.1
decorator==4.4.0
Django==2.2
django-extensions==2.1.6
Fabric3==1.14.post1
gcg==0.1.8
gitdb2==2.0.5
GitPython==2.1.8
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
paramiko==2.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pyasn1==0.4.5
pycparser==2.19
Pygments==2.3.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pytz==2018.9
Send2Trash==1.5.0
six==1.12.0
smmap2==2.0.5
sqlparse==0.3.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.15.2
